AWS/S3/requirements.txt
boto3   # aws s3 통신용
python-multipart   # 파일 업로드
pymysql    # RDS
sqlalchemy # RDS
uvicorn 
fastapi 
python-dotenv